Coup Fourre should leave player with ROLL !
Coup Fourre is supposed to cancel out the hazard card completely,
as if it had never been played. It gets placed in the discard pile.
The card under the hazard card will always have been a "ROLL !" card,
since it's not legal to play a hazard card on anything else.